requirements.txt
Django==3.0.1
djangorestframework==3.11.0
gunicorn==20.0.4
Pillow==7.0.0
postgres==3.0.0
psycopg2-binary==2.8.5
psycopg2-pool==1.1
ptyprocess==0.6.0
Pygments==2.6.1
requests==2.22.0
requests-oauthlib==1.3.0
sqlparse==0.3.0
pytz==2019.3
sqlparse==0.3.1
whitenoise==5.0.1
asgiref==3.2.7